requirements.txt
amqp==1.4.6
anyjson==0.3.3
boto==2.38.0
bottle==0.12.4
configlib==1.0.0
configparser==3.5.0b2
docutils==0.11
elasticutils==0.8.2
futures==2.1.6
glob2==0.4.1
kombu==3.0.26
pyelasticsearch==0.6.1
pyes==0.99.5
pymongo==2.6.3
python-dateutil==2.2
pytz==2013.9
requests==2.5.3
requests-futures==0.9.5
s3cmd==1.0.1
simplejson==3.3.3
six==1.5.2
urllib3==1.7.1
virtualenv==1.11.4
wsgiref==0.1.2
kitnirc==0.2.6
netaddr==0.7.11
pygeoip==0.3.1
pynsive==0.2.6
MySQL-python==1.2.5
ipaddr==2.1.11
ipwhois==0.9.0
celery==3.1.17
#only needed for google integration to pull gdocs/mail logs
oauth2client==1.4.7
google-api-python-client==1.4.0
pytx==0.1.3
